#377 - refactor delete sql generator
