+++ src/main/java/com/homihq/db2rest/jdbc/config/JdbcSqlCreatorConfiguration.java
@@ -2,8 +2,6 @@ package com.homihq.db2rest.jdbc.config;
 
 
 import com.homihq.db2rest.core.Dialect;
-import com.homihq.db2rest.jdbc.sql.CreateCreatorTemplate;
-import com.homihq.db2rest.jdbc.sql.DeleteCreatorTemplate;
 import com.homihq.db2rest.jdbc.sql.UpdateCreatorTemplate;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.context.annotation.Bean;
@@ -17,20 +15,12 @@ import javax.sql.DataSource;
 @ConditionalOnBean(DataSource.class)
 public class JdbcSqlCreatorConfiguration {
 
-    @Bean
-    public CreateCreatorTemplate createCreatorTemplate(SpringTemplateEngine templateEngine) {
-        return new CreateCreatorTemplate(templateEngine);
-    }
 
     @Bean
     public UpdateCreatorTemplate updateCreatorTemplate(SpringTemplateEngine templateEngine, Dialect dialect) {
         return new UpdateCreatorTemplate(templateEngine, dialect);
     }
 
-    @Bean
-    public DeleteCreatorTemplate deleteCreatorTemplate(SpringTemplateEngine templateEngine, Dialect dialect) {
-        return new DeleteCreatorTemplate(templateEngine, dialect);
-    }
 
 
 }
