+++ src/main/java/com/homihq/db2rest/jdbc/service/JdbcCreateService.java
@@ -6,7 +6,7 @@ import com.homihq.db2rest.core.service.CreateService;
 import com.homihq.db2rest.exception.GenericDataAccessException;
 import com.homihq.db2rest.core.model.DbColumn;
 import com.homihq.db2rest.core.model.DbTable;
-import com.homihq.db2rest.jdbc.sql.CreateCreatorTemplate;
+import com.homihq.db2rest.jdbc.sql.SqlCreatorTemplate;
 import com.homihq.db2rest.rest.create.dto.CreateContext;
 import com.homihq.db2rest.rest.create.dto.CreateResponse;
 import com.homihq.db2rest.jdbc.tsid.TSIDProcessor;
@@ -29,7 +29,7 @@ import static org.springframework.util.CollectionUtils.isEmpty;
 public class JdbcCreateService implements CreateService {
 
     private final TSIDProcessor tsidProcessor;
-    private final CreateCreatorTemplate createCreatorTemplate;
+    private final SqlCreatorTemplate sqlCreatorTemplate;
     private final SchemaCache schemaCache;
     private final DbOperationService dbOperationService;
     private final Dialect dialect;
@@ -62,7 +62,7 @@ public class JdbcCreateService implements CreateService {
             this.dialect.processTypes(dbTable, insertableColumns, data);
 
             CreateContext context = new CreateContext(dbTable, insertableColumns);
-            String sql = createCreatorTemplate.createQuery(context);
+            String sql = sqlCreatorTemplate.createQuery(context);
 
 
             log.debug("SQL - {}", sql);
