+++ src/main/java/com/homihq/db2rest/jdbc/sql/SqlCreatorTemplate.java
@@ -1,7 +1,10 @@
 package com.homihq.db2rest.jdbc.sql;
 
+import com.homihq.db2rest.core.Dialect;
 import com.homihq.db2rest.core.model.DbColumn;
 import com.homihq.db2rest.core.model.DbSort;
+import com.homihq.db2rest.rest.create.dto.CreateContext;
+import com.homihq.db2rest.rest.delete.dto.DeleteContext;
 import com.homihq.db2rest.rest.read.dto.ReadContext;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -21,6 +24,36 @@ public class SqlCreatorTemplate {
 
 
     private final SpringTemplateEngine templateEngine;
+    private final Dialect dialect;
+    public String deleteQuery(DeleteContext deleteContext) {
+
+        String rendererTableName = dialect.supportAlias() ? deleteContext.getTable().render()
+                : deleteContext.getTable().name();
+
+        Map<String,Object> data = new HashMap<>();
+        data.put("rootTable", rendererTableName);
+        data.put("rootWhere", deleteContext.getWhere());
+
+        Context context = new Context();
+        context.setVariables(data);
+        return templateEngine.process("delete", context);
+
+    }
+
+    public String createQuery(CreateContext createContext) {
+
+        Map<String,Object> data = new HashMap<>();
+
+        data.put("table", createContext.table().fullName());
+        data.put("columns", createContext.renderColumns());
+        data.put("parameters", createContext.renderParams());
+
+
+        Context context = new Context();
+        context.setVariables(data);
+        return templateEngine.process("insert", context);
+
+    }
 
     public String findOne(ReadContext readContext) {
         Map<String,Object> data = new HashMap<>();
